Attempting to change how the navigation links work in responsive mode to not hide the main content.
