+++ src/app/App.tsx
@@ -1,4 +1,3 @@
-import {useState} from "react";
 import './App.css'
 import Navigation from "../components/navigation/navigation.tsx"
 import Hero from "../components/hero/hero.tsx"
@@ -8,20 +7,10 @@ import Footer from "../components/footer/footer.tsx"
 
 const App = ()=> {
 
-    const [showNavList, setShowNavList] = useState(false)
-
-    const displayNavList = ()=> {
-        setShowNavList(true);
-    }
-
-    const hideNavList = () => {
-        setShowNavList(false);
-    }
-
     return (
         <>
-            <Navigation showNavList={showNavList} displayNavList={displayNavList} hideNavList={hideNavList} />
-            <section className={`mainSection ${showNavList ? "displayNav" : ""}`}>
+            <Navigation />
+            <section className={"mainSection"}>
                 <Hero />
                 <Projects />
                 <Testimonials />
